+++ MonkeyDungeon/GameFeatures/ScalingValue.cs
@@ -8,14 +8,14 @@ namespace MonkeyDungeon.GameFeatures
 {
     public class ScalingValue
     {
-        public readonly float Inital_Value;
-        public float Value { get; private set; }
-        public float Scaling_Rate { get; private set; }
+        public readonly double Inital_Value;
+        public double Value { get; private set; }
+        public double Scaling_Rate { get; private set; }
         public void Change_ScalingRate(float rate) { Scaling_Rate = rate; Adjust(); }
         public int Scaling_Level { get; private set; }
         public void Change_ScalingLevel(int level) { Scaling_Level = level; Adjust(); }
 
-        public ScalingValue(float initalValue, int level, float rate)
+        public ScalingValue(double initalValue, int level, double rate)
         {
             Inital_Value = initalValue;
             Value = initalValue;
@@ -29,6 +29,6 @@ namespace MonkeyDungeon.GameFeatures
             Value = (Scaling_Level * Scaling_Rate) + Inital_Value;
         }
 
-        public static implicit operator float(ScalingValue sv) => sv.Value;
+        public static implicit operator double(ScalingValue sv) => sv.Value;
     }
 }
